requirements.txt
streamlit
openai
huggingface_hub
pillow
